requirements.txt
streamlit==1.25.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
python-dotenv==1.0.0
google-generativeai==0.5.0